src/backend/integrity_service/app/services/timestamp_service.py
"""
RFC 3161 Timestamping Service for ACGS-PGP Framework
Implements trusted timestamping for audit logs and policy rules
"""

import hashlib
import base64
import requests
from datetime import datetime, timezone
from typing import Optional, Dict, Any, Tuple
import logging

logger = logging.getLogger(__name__)


class RFC3161TimestampService:
    """
    RFC 3161 Timestamp Authority (TSA) client implementation
    """
    
    def __init__(self, tsa_url: str = "http://timestamp.digicert.com", timeout: int = 30):
        """
        Initialize timestamp service
        
        Args:
            tsa_url: Timestamp Authority URL
            timeout: Request timeout in seconds
        """
        self.tsa_url = tsa_url
        self.timeout = timeout
        self.hash_algorithm = "SHA3-256"
        
    def create_timestamp_request(self, message_hash: bytes, nonce: Optional[int] = None) -> bytes:
        """
        Create RFC 3161 timestamp request (TSRequest)
        
        Args:
            message_hash: Hash of the message to timestamp
            nonce: Optional nonce for request uniqueness
            
        Returns:
            ASN.1 DER encoded timestamp request
        """
        # This is a simplified implementation
        # In production, use proper ASN.1 library like pyasn1
        
        # For now, create a basic request structure
        # Real implementation would use proper ASN.1 encoding
        request_data = {
            "version": 1,
            "messageImprint": {
                "hashAlgorithm": self.hash_algorithm,
                "hashedMessage": base64.b64encode(message_hash).decode('ascii')
            },
            "reqPolicy": None,
            "nonce": nonce,
            "certReq": True,
            "extensions": None
        }
        
        # Convert to bytes (simplified - real implementation needs ASN.1)
        request_json = str(request_data).encode('utf-8')
        return request_json
    
    def send_timestamp_request(self, message_hash: bytes) -> Optional[Dict[str, Any]]:
        """
        Send timestamp request to TSA and get response
        
        Args:
            message_hash: Hash of the message to timestamp
            
        Returns:
            Dictionary containing timestamp response data or None if failed
        """
        try:
            # Create timestamp request
            ts_request = self.create_timestamp_request(message_hash)
            
            # Send request to TSA
            headers = {
                'Content-Type': 'application/timestamp-query',
                'Content-Length': str(len(ts_request))
            }
            
            response = requests.post(
                self.tsa_url,
                data=ts_request,
                headers=headers,
                timeout=self.timeout
            )
            
            if response.status_code == 200:
                # Parse timestamp response (simplified)
                timestamp_token = response.content
                
                # Extract timestamp value (simplified - real implementation needs ASN.1 parsing)
                timestamp_value = datetime.now(timezone.utc)
                
                result = {
                    "timestamp_token": timestamp_token,
                    "timestamp_value": timestamp_value,
                    "tsa_url": self.tsa_url,
                    "hash_algorithm": self.hash_algorithm,
                    "message_imprint": message_hash,
                    "status": "success"
                }
                
                logger.info(f"Timestamp request successful: {timestamp_value}")
                return result
            else:
                logger.error(f"Timestamp request failed: {response.status_code}")
                return None
                
        except requests.RequestException as e:
            logger.error(f"Timestamp request error: {e}")
            return None
        except Exception as e:
            logger.error(f"Unexpected error in timestamp request: {e}")
            return None
    
    def verify_timestamp_token(self, timestamp_token: bytes, message_hash: bytes) -> bool:
        """
        Verify RFC 3161 timestamp token
        
        Args:
            timestamp_token: ASN.1 DER encoded timestamp token
            message_hash: Original message hash
            
        Returns:
            True if timestamp is valid, False otherwise
        """
        try:
            # Simplified verification - real implementation needs:
            # 1. ASN.1 parsing of timestamp token
            # 2. Certificate chain validation
            # 3. Signature verification
            # 4. Message imprint verification
            
            # For now, basic validation
            if not timestamp_token or len(timestamp_token) < 10:
                return False
            
            # Check if message hash matches (simplified)
            # Real implementation would extract and compare message imprint from token
            
            logger.debug("Timestamp token verification (simplified)")
            return True
            
        except Exception as e:
            logger.error(f"Timestamp verification error: {e}")
            return False
    
    def extract_timestamp_value(self, timestamp_token: bytes) -> Optional[datetime]:
        """
        Extract timestamp value from RFC 3161 token
        
        Args:
            timestamp_token: ASN.1 DER encoded timestamp token
            
        Returns:
            Extracted timestamp or None if failed
        """
        try:
            # Simplified extraction - real implementation needs ASN.1 parsing
            # For now, return current time as placeholder
            return datetime.now(timezone.utc)
            
        except Exception as e:
            logger.error(f"Timestamp extraction error: {e}")
            return None


class MockTimestampService(RFC3161TimestampService):
    """
    Mock timestamp service for testing and development
    """
    
    def __init__(self):
        super().__init__(tsa_url="mock://localhost")
        
    def send_timestamp_request(self, message_hash: bytes) -> Optional[Dict[str, Any]]:
        """
        Mock timestamp request that always succeeds
        """
        try:
            # Create mock timestamp token
            timestamp_value = datetime.now(timezone.utc)
            mock_token = f"MOCK_TIMESTAMP_{timestamp_value.isoformat()}_{base64.b64encode(message_hash).decode()}"
            timestamp_token = mock_token.encode('utf-8')
            
            result = {
                "timestamp_token": timestamp_token,
                "timestamp_value": timestamp_value,
                "tsa_url": self.tsa_url,
                "hash_algorithm": self.hash_algorithm,
                "message_imprint": message_hash,
                "status": "success",
                "mock": True
            }
            
            logger.info(f"Mock timestamp created: {timestamp_value}")
            return result
            
        except Exception as e:
            logger.error(f"Mock timestamp error: {e}")
            return None
    
    def verify_timestamp_token(self, timestamp_token: bytes, message_hash: bytes) -> bool:
        """
        Mock timestamp verification
        """
        try:
            token_str = timestamp_token.decode('utf-8')
            return token_str.startswith("MOCK_TIMESTAMP_")
        except:
            return False


class TimestampManager:
    """
    High-level timestamp management service
    """
    
    def __init__(self, use_mock: bool = True):
        """
        Initialize timestamp manager
        
        Args:
            use_mock: Whether to use mock service for development
        """
        if use_mock:
            self.timestamp_service = MockTimestampService()
        else:
            # Use real TSA services
            self.timestamp_service = RFC3161TimestampService()
        
        self.use_mock = use_mock
    
    def timestamp_data(self, data: str) -> Optional[Dict[str, Any]]:
        """
        Create timestamp for arbitrary data
        
        Args:
            data: Data to timestamp
            
        Returns:
            Timestamp result dictionary or None if failed
        """
        # Generate hash of data
        message_hash = hashlib.sha3_256(data.encode('utf-8')).digest()
        
        # Get timestamp
        result = self.timestamp_service.send_timestamp_request(message_hash)
        
        if result:
            result["original_data_hash"] = message_hash.hex()
            
        return result
    
    def timestamp_audit_log(self, log_entry: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """
        Create timestamp for audit log entry
        
        Args:
            log_entry: Audit log entry dictionary
            
        Returns:
            Timestamp result dictionary or None if failed
        """
        # Create deterministic representation of log entry
        import json
        log_json = json.dumps(log_entry, sort_keys=True, separators=(',', ':'))
        
        return self.timestamp_data(log_json)
    
    def timestamp_policy_rule(self, rule_content: str, metadata: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """
        Create timestamp for policy rule
        
        Args:
            rule_content: Policy rule content
            metadata: Rule metadata
            
        Returns:
            Timestamp result dictionary or None if failed
        """
        # Combine rule content and metadata
        combined_data = {
            "rule_content": rule_content,
            "metadata": metadata
        }
        
        import json
        combined_json = json.dumps(combined_data, sort_keys=True, separators=(',', ':'))
        
        return self.timestamp_data(combined_json)
    
    def verify_timestamp(self, timestamp_token: bytes, original_data: str) -> bool:
        """
        Verify timestamp token against original data
        
        Args:
            timestamp_token: RFC 3161 timestamp token
            original_data: Original data that was timestamped
            
        Returns:
            True if timestamp is valid, False otherwise
        """
        # Generate hash of original data
        message_hash = hashlib.sha3_256(original_data.encode('utf-8')).digest()
        
        # Verify timestamp
        return self.timestamp_service.verify_timestamp_token(timestamp_token, message_hash)


# Global timestamp manager instance
timestamp_manager = TimestampManager(use_mock=True)  # Use mock for development
