requirements-test.txt
# ACGS-PGP Testing Dependencies
# Enhanced dependency mocking and testing framework

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0                    # Enhanced mocking capabilities
pytest-cov==4.1.0                      # Coverage reporting
pytest-xdist==3.3.1                    # Parallel test execution
pytest-timeout==2.1.0                  # Test timeout handling
pytest-benchmark==5.1.0                # Performance benchmarking
pytest-html==4.1.1                     # HTML test reports
pytest-json-report==1.5.0              # JSON test reports

# HTTP and API testing
httpx==0.25.2                          # Async HTTP client for testing
responses==0.24.1                      # HTTP request mocking
aioresponses==0.7.4                    # Async HTTP mocking
pytest-httpserver==1.1.3               # Local HTTP server for testing
requests-mock==1.11.0                  # Requests library mocking

# Database testing
pytest-postgresql==5.0.0               # PostgreSQL testing utilities
pytest-alembic==0.12.1                 # Alembic migration testing
sqlalchemy-utils==0.41.1               # Database testing utilities
factory-boy==3.3.0                     # Test data factories

# Async testing utilities
asynctest==0.13.0                      # Async testing utilities
pytest-trio==0.8.0                     # Trio async testing
aiofiles==23.2.1                       # Async file operations for testing

# Mock and fixture utilities
freezegun==1.2.2                       # Time mocking
faker==20.1.0                          # Fake data generation
mimesis==11.1.0                        # Alternative fake data generator
pyfakefs==5.3.2                        # Filesystem mocking

# Security testing
pytest-security==0.1.1                 # Security testing utilities
bandit==1.7.5                          # Security linting

# Performance and load testing
locust==2.37.0                         # Load testing framework
memory-profiler==0.61.0                # Memory usage profiling

# Validation and schema testing
jsonschema==4.24.0                     # JSON schema validation                   # JSON schema validation
cerberus==1.3.7                        # Data validation
hypothesis==6.88.1                     # Property-based testing

# Logging and debugging
pytest-logging==2016.11.4              # Enhanced logging for tests
pytest-clarity==1.0.1                  # Better assertion output
pytest-sugar==0.9.7                    # Better test output formatting

# Integration testing utilities
docker==6.1.3                          # Docker integration for tests
testcontainers==4.10.0                  # Container-based testing
wiremock==2.6.1                        # Service virtualization

# Cryptographic testing
cryptography==44.0.1                   # Cryptographic testing utilities
pycryptodome==3.23.0                   # Additional crypto testing

# LLM and AI testing mocks
openai==1.3.7                          # OpenAI API mocking
anthropic==0.7.7                       # Anthropic API mocking

# Shared dependencies (ensure compatibility)
fastapi==0.104.1
pydantic==2.11.5
sqlalchemy==2.0.41
alembic==1.16.1
python-dotenv==1.1.0
numpy==2.2.6
tenacity==8.2.3                           # Retry logic for LLM integration
